backend/requirements.txt
asgiref==3.8.1
attrs==20.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cs50==9.4.0
Django==5.2
django-emoji==2.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
jellyfish==0.11.2
Jinja2==3.1.6
lib50==3.0.12
MarkupSafe==3.0.2
maturin==1.8.3
msgspec==0.19.0
packaging==24.2
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==78.1.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
submit50==3.2.0
termcolor==1.1.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.10.0
